stringmatching renaming left overs
